requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
flask-restx==1.2.0
pyjwt==2.8.0